feat(121684): :art: altera local do handleFinalizarMedicao para poder reutilizar
